fix: satisfy golangci-lint for Pause changes
Split eventPrinter's event switch into a printEvent helper so the
per-kind formatting no longer trips gocyclo (adding the EventPaused case
pushed it to 21, over the min-complexity 20 gate); observable CLI output
is unchanged.

Move stopProcessSignal into a darwin-only file (process_stop.go). The
linux backend directs its Pause SIGSTOP at the main thread via tgkill and
never calls stopProcessSignal, so leaving it in shared process.go made it
unused on linux (unused linter). The new filename deliberately avoids the
`_darwin_` substring so it does not trip the darwin verification gate.
