+++ internal/debugger/process.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
-	"syscall"
 )
 
 // process tracks the OS handle for the tracee, with platform-specific hooks
@@ -61,28 +60,6 @@ func (p *process) kill(b Backend) error {
 	p.live = false
 	if err := killProcess(b, p.pid, p.cmd); err != nil {
 		return fmt.Errorf("kill: %w", err)
-	}
-	return nil
-}
-
-// stopProcessSignal sends a whole-thread-group SIGSTOP to pid. It is the
-// darwin backend's StopProcess() mechanism (the linux backend overrides
-// StopProcess to direct the signal at the main thread via tgkill; see
-// backend_linux_amd64.go for why). syscall.Kill is used directly instead of
-// os.FindProcess(pid).Signal: os.FindProcess never fails to find a process on
-// Unix (it just wraps the pid), so that pairing never actually distinguished
-// "no such process" from any other error. ESRCH (process already gone) is
-// treated as an idempotent no-op, matching process.kill's idempotency above.
-// NOTE: on darwin this is Pause *groundwork* only — macOS delivers a SIGSTOP to
-// a ptraced tracee as a job-control stop that our wait4 loop never reports, so
-// it does not yet surface as EventPaused. A functional darwin Pause needs Mach
-// thread_suspend instead; see AGENTS.md → Pause (darwin caveat).
-func stopProcessSignal(pid int) error {
-	if pid == 0 {
-		return fmt.Errorf("StopProcess: no process")
-	}
-	if err := syscall.Kill(pid, syscall.SIGSTOP); err != nil && err != syscall.ESRCH {
-		return fmt.Errorf("StopProcess: %w", err)
 	}
 	return nil
 }
